### 2022-12-15 #564 @RWParsons
Presubmission inquiry - predictNMB: Evaluate Clinical Prediction Models by Net Monetary Benefit
Submitting Author Name: Rex Parsons
Submitting Author Github Handle: @RWParsons
Other Package Authors Github handles: @robinblythe, @agbarnett
Repository:  https://github.com/RWParsons/predictNMB
Submission type: Pre-submission

---

```
Package: predictNMB
Title: Evaluate Clinical Prediction Models by Net Monetary Benefit
Version: 0.0.0.9000
    person("Rex", "Parsons", role = c("aut", "cre"),
           email = "rex.parsons94@gmail.com",
    person("Robin", "Blythe", role = "aut",
    person("Adrian", "Barnett", role = "aut",
    person("Susanna", "Cramb", role = "ctb",
    person("Steve", "McPhail", role = "ctb",
           comment = c(ORCID = "0000-0002-1463-662X"))
  )
Description: Estimates when and where a model-guided treatment strategy may outperform
    a treat-all or treat-none approach by Monte Carlo simulation and evaluation of the Net Monetary Benefit (NMB).
    assertthat,
    cutpointr,
    pmsampsize,
    tidyr
    flextable,
    parallel,

```

The package helps the user perform Monte Carlo simulations to evaluate the performance of a (simulated) clinical prediction model to assign (simulated) patients to treatment options with user-defined levels of performance/costs to evaluate the costs/benefits of the use of that model-decision process. I could also consider this a wrapper package as it really just abstracts away the complexity of performing this Monte Carlo simulation study from the user rather than adding anything particularly fancy regarding the Monte Carlo procedure.

Yes.

People considering the development of a clinical decision support system used to guide patient care between two possible interventions (or non-treatment) and wanting to know how likely it is that their model will be clinically useful compared to a treat-all or treat-none strategy. Performing simulations using hypothetical clinical prediction models of likely estimated performance may show that value is unlikely and deter researchers from model development efforts. These models are rarely used in clinical practice due to them often failing to improve patient outcomes compared to standard care. All evaluation is done in health economic terms, to orient the results towards value-based care.

To the best of my knowledge, there are no other packages to simulate prediction models and streamline their evaluation of them in monetary terms.

Yes.

Much of the standards from srr are irrelevant for this package since all data analysed are simulated (I checked the general and bayesian standards).
In terms of G1.0 (*Statistical Software should list at least one primary reference from published academic literature.*), I currently have a journal article under review which uses the simulation process here for another purpose but there are no other existing literature which describe this process alongside this form of evaluation (as far as I know). The literature that we cite in the paper to create our simulated models is cited in the introduction-to-predictNMB vignette.
