### 2022-12-15 8ea16c2
add codecov to actions and badge to README
