### 2022-12-15 a9e4d28
add tests for print and more table maker
