### 2022-12-15 #564 @annakrystalli
Hello @RWParsons 👋

Many thanks for your pre-submission enquiry!

The editorial team has concluded that the package definitely fits in our "stats" scope.

To confirm everything is okay however, before moving to full submission, we do suggest going through the formal process of documenting compliance with the stats standards. Although you've stated that standards have been documented, the `srr` check indicates that they have not. It would be great if you could complete them to ensure submission would be accepted.

You can call `@ropensci-review-bot check srr` yourself in this issue at any time to confirm documentation has been completed successfully. You can find [more details in our documentation](https://stats-devguide.ropensci.org/pkgdev.html#pkgdev-srr).

Any questions, just let me know!
