+++ man/print.predictNMBscreen.Rd
@@ -12,14 +12,14 @@
 \item{...}{Optional, ignored arguments.}
 }
 \value{
-`print` returns `x` invisibly.
+\code{print} returns \code{x} invisibly.
 if (FALSE) {
-  get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
-  sim_screen_obj <- screen_simulation_inputs(
-    n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1), event_rate = 0.1,
-    fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
-  )
-  print(sim_screen_obj)
+get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
+sim_screen_obj <- screen_simulation_inputs(
+n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1), event_rate = 0.1,
+fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+)
+print(sim_screen_obj)
 }
 }
 \description{
