### 2022-12-15 1d417d6
remove \'s before %'s that's now unnecessary
the addition of srr in previous commit means that roxygen doesn't need the %'s escaped but does this automatically when running document(). This commit removes the escapes in the roxygen sections for function documentation as they were becoming \\ and causing problems again.
