+++ .github/workflows/test-coverage.yaml
@@ -31,7 +31,8 @@ jobs:
           covr::codecov(
             quiet = FALSE,
             clean = FALSE,
-            install_path = file.path(Sys.getenv("RUNNER_TEMP"), "package")
+            install_path = file.path(Sys.getenv("RUNNER_TEMP"), "package"),
+            token = "${{ secrets.COVERALLS_TOKEN }}"
           )
         shell: Rscript {0}
 
