+++ .github/workflows/test-coverage.yaml
@@ -32,7 +32,7 @@ jobs:
             quiet = FALSE,
             clean = FALSE,
             install_path = file.path(Sys.getenv("RUNNER_TEMP"), "package"),
-            token = "${{ secrets.COVERALLS_TOKEN }}"
+            token = "${{ secrets.CODECOV_TOKEN }}"
           )
         shell: Rscript {0}
 
