### 2022-12-17 9fa1089
add error message for inappropriate input lengths
