### 2022-12-18 8c7abe7
add test for catching bad inputs error
