### 2022-12-18 1c3fc4d
Create test-simulate-calibrated-model.R
