### 2022-12-18 1bf4d38
use fixed random seed for test
