### 2022-12-18 84fb53d
test for zero length input and repeatability across seeds
