### 2022-12-18 cc66fda
document ssr throughout package
