+++ R/do_nmb_sim.R
@@ -62,8 +62,9 @@
 #' \cr\cr
 #' \code{get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)}
 #'
-#'
 #' There is a helper function, \code{get_nmb_sampler()}, to help you create these.
+#'
+#' @srrstats {G2.1a} Data types for all inputs are documented.
 #'
 #' @return Returns a \code{predictNMBsim} object.
 #' @export
