+++ tests/testthat.R
@@ -1,3 +1,6 @@
+#' @srrstats {G5.2, G5.2a, G5.2b, G5.3, G5.4, G5.4a, G5.5, G5.6, G5.6a, G5.6b, G5.8, G5.8a, G5.8b, G5.8c, G5.8d, G5.9, G5.11}
+#' Covered by package tests.
+
 library(testthat)
 library(predictNMB)
 
