+++ R/predictNMB-package.R
@@ -1,0 +1,16 @@
+#' @description
+#' To learn more about how to use predictNMB, see the vignettes at the
+#' \href{https://rwparsons.github.io/predictNMB/index.html}{website} or using the following code:
+#'
+#' `browseVignettes(package = "predictNMB")`
+#' @keywords internal
+#' @srrstats {G1.3} All statistical terminology is defined and explained.
+#' @srrstats {G1.4, G1.4a} Roxygen is used for all documentation.
+#' @srrstats {G1.5, G1.5} No performance claims are made.
+#' @srrstats {G2.15} Functions never assume non-missingness.
+#'
+"_PACKAGE"
+
+## usethis namespace: start
+## usethis namespace: end
+NULL
