+++ tests/testthat/test-simulate-calibrated-model.R
@@ -1,5 +1,4 @@
 test_that("Simulated model has approximately right AUC and is well calibrated", {
-
   input_auc <- 0.75
 
   withr::with_seed(
@@ -16,14 +15,14 @@ test_that("Simulated model has approximately right AUC and is well calibrated", 
   training_data$preds <- stats::predict(model, type = "response")
 
   x <- expand.grid(
-    pos = training_data[training_data$actual==1, "preds"],
-    neg = training_data[training_data$actual==0, "preds"]
+    pos = training_data[training_data$actual == 1, "preds"],
+    neg = training_data[training_data$actual == 0, "preds"]
   )
 
   output_auc <- mean(x$pos > x$neg)
 
   # AUC is close to what was input to get_sample()
-  expect_true(output_auc > input_auc*0.95, output_auc < input_auc * 1.05)
+  expect_true(output_auc > input_auc * 0.95, output_auc < input_auc * 1.05)
 
   calibration_model <- stats::lm(
     actual ~ preds,
@@ -34,9 +33,9 @@ test_that("Simulated model has approximately right AUC and is well calibrated", 
 
   # model is well calibrated
   expect_true(
-      conf.ints[1,1] < 0 &
-      conf.ints[1,2] > 0 &
-      conf.ints[2,1] < 1 &
-      conf.ints[2,2] > 1
+    conf.ints[1, 1] < 0 &
+      conf.ints[1, 2] > 0 &
+      conf.ints[2, 1] < 1 &
+      conf.ints[2, 2] > 1
   )
 })
