### 2022-12-18 004ea61
use linewidth vs size arg based on ggplot2 version
