### 2022-12-18 #564 @RWParsons
Hi @annakrystalli!
Thanks for the quick response and sorry for the rookie mistake. I've updated the documentation for `srr` but currently only for the general stats requirements. I know you said it fits within the stats scope but I'm just wondering whether that'd be within the "Bayesian and Monte Carlo Routines" subsection or whether it's better suited elsewhere (maybe "EDA" instead just based on the standards included?).
