### 2022-12-19 #564 @annakrystalli
Hello @RWParsons, good question.

This is where the act of documenting against standards can be helpful in narrowing down the category. The [stats-devguide states categories are appropriate where at least half of all standards can be applied](https://stats-devguide.ropensci.org/pkgdev.html#scope). Given that and your efforts to add the standard documentation, have they made you lean more towards one or the other categories? Would atleast half the Bayesian & Monte Carlo routines standards apply?
