### 2022-12-20 #564 @RWParsons
Thanks for the quick response and advice @annakrystalli !
I've just tallied up how many would be NA for the Bayesian and the EDA standards and I don't make the cut for the Bayesian standards (only get about 1/3 to be relevant) but I do for the EDA (29/34 standards are relevant).

I'll make the changes to the package and document the EDA standards throughout before calling the srr checker bot.

Thanks!
