### 2022-12-20 03e7bd0
create index for joins in screen_simulation_inputs()
