+++ R/screen_simulation_inputs.R
@@ -38,6 +38,7 @@
 #' @srrstats {G2.1a} Data types for all inputs are documented.
 #' @srrstats {G2.8} Appropriate conversions applied before passing of inputs.
 #' @srrstats {G2.11} tibbles used for appropriate processing of columns of functions.
+#' @srrstats {EA2.0, EA2.1, EA2.2, EA2.2a, EA2.2b} Plotting of 'predictNMBscreen' objects uses the index column from \code{input_grid} to perform joins. All values within the input_grid are unique to the combination of possible input values (created in \code{screen_simulation_inputs()} using \code{.sim_id = dplyr::row_number()}). Attribute is set for index column name.
 #'
 #' @return Returns a \code{predictNMBscreen} object.
 #' @export
@@ -128,7 +129,8 @@ screen_simulation_inputs <- function(sample_size, n_sims, n_valid, sim_auc, even
   input_grid <-
     input_grid %>%
     dplyr::inner_join(small_grid, by = "small_grid_id") %>%
-    dplyr::select(-small_grid_id)
+    dplyr::select(-small_grid_id) %>%
+    dplyr::mutate(.sim_id = dplyr::row_number())
 
   summary_grid <- input_grid %>%
     dplyr::mutate(
@@ -200,6 +202,8 @@ screen_simulation_inputs <- function(sample_size, n_sims, n_valid, sim_auc, even
   )
 
   class(res) <- "predictNMBscreen"
+
+  attr(res, "index") <- ".sim_id"
 
   res
 }
