### 2022-12-20 db1b1f9
make_summary_table() returns tibble for screen
and srr docs
