### 2022-12-20 abaa857
update plot tests and move to test-plot.R
