+++ DESCRIPTION
@@ -40,6 +40,7 @@ Suggests:
     parallel,
     rmarkdown,
     testthat (>= 3.0.0),
+    vdiffr,
     withr
 Config/testthat/edition: 3
 VignetteBuilder: knitr
