### 2022-12-20 34ef0f4
update make_summary_table() documentation
