+++ R/make_summary_table.R
@@ -11,6 +11,9 @@
 #' @param rename_vector A named vector for renaming the methods in the summary.
 #' The values of the vector are the default names and the names given are the
 #' desired names in the output.
+#' @param show_full_inputs A logical. Whether or not to include the inputs used
+#' for simulation alongside aggregations. (Only applicable for
+#' \code{predictNMBscreen} objects created using \code{screen_simulation_inputs()})
 #' @param ... Additional, ignored arguments.
 #'
 #' @details Table summaries will be based on the `what` argument.
@@ -23,7 +26,7 @@
 #' @srrstats {EA5.2} There is no typical onscreen printing from \code{make_summary_table()}. It instead returns a \code{tbl_df/tbl/data.frame} with aggregate values. The default is to round the values but the user is able to overwrite this by using the \code{agg_functions} argument and pass the output to a more general table formatting package like \code{flextable}.
 #' @srrstats {EA5.3} Summary statsistics indicate \code{storage.mode} via tibble.
 #'
-#' @return Returns a \code{data.frame}.
+#' @return Returns a \code{tibble}.
 #' @export
 #' @examples
 #'
@@ -45,10 +48,25 @@ make_summary_table <- function(x,
                                  "95% CI" = function(x) paste0(round(stats::quantile(x, probs = c(0.025, 0.975)), digits = 1), collapse = " to ")
                                ),
                                rename_vector,
+                               show_full_inputs = FALSE,
                                ...) {
   UseMethod("make_summary_table")
 }
 
+
+
+
+#' Gets the relevant simulation data for summarising in \code{make_summary_table()}
+#'
+#' @param x A \code{predictNMBsim} object.
+#' @param what What to summarise: one of "nmb", "inb" or "cutpoints".
+#' Defaults to "nmb".
+#' @param inb_ref_col Which cutpoint method to use as the reference strategy
+#' when calculating the incremental net monetary benefit. See \code{do_nmb_sim}
+#' for more information.
+#'
+#' @return A \code{data.frame}
+#' @noRd
 get_sim_data <- function(x, what, inb_ref_col = NULL) {
   if (what %in% c("nmb", "inb")) {
     res <- x$df_result
@@ -67,7 +85,42 @@ get_sim_data <- function(x, what, inb_ref_col = NULL) {
 }
 
 
+#' Create table summaries of \code{predictNMBscreen} objects.
+#'
+#' @param x A \code{predictNMBscreen} object.
+#' @param what What to summarise: one of "nmb", "inb" or "cutpoints".
+#' Defaults to "nmb".
+#' @param inb_ref_col Which cutpoint method to use as the reference strategy
+#' when calculating the incremental net monetary benefit. See \code{do_nmb_sim}
+#' for more information.
+#' @param agg_functions A named list of functions to use to aggregate the
+#' selected values. Defaults to the median and 95% interval.
+#' @param rename_vector A named vector for renaming the methods in the summary.
+#' The values of the vector are the default names and the names given are the
+#' desired names in the output.
+#' @param show_full_inputs A logical. Whether or not to include the inputs used
+#' for simulation alongside aggregations.
+#' @param ... Additional, ignored arguments.
+#'
+#' @details Table summaries will be based on the `what` argument.
+#' Using "nmb" returns the simulated values for NMB, with no reference group;
+#' "inb" returns the difference between simulated values for NMB and a set
+#' strategy defined by `inb_ref_col`; "cutpoints" returns the cutpoints
+#' selected (0, 1).
+#'
[…]
