+++ man/make_summary_table.predictNMBsim.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/make_summary_table.R
-\name{make_summary_table}
-\alias{make_summary_table}
-\title{Create table summaries of simulation objects.}
+\name{make_summary_table.predictNMBsim}
+\alias{make_summary_table.predictNMBsim}
+\title{Create table summaries of \code{predictNMBsim} objects.}
 \usage{
-make_summary_table(
+\method{make_summary_table}{predictNMBsim}(
   x,
   what = c("nmb", "inb", "cutpoints"),
   inb_ref_col = NULL,
@@ -12,12 +12,11 @@ make_summary_table(
     `95\% CI` = function(x) paste0(round(stats::quantile(x, probs = c(0.025, 0.975)),
     digits = 1), collapse = " to ")),
   rename_vector,
-  show_full_inputs = FALSE,
   ...
 )
 }
 \arguments{
-\item{x}{A \code{predictNMBscreen} or \code{predictNMBsim} object.}
+\item{x}{A \code{predictNMBsim} object.}
 
 \item{what}{What to summarise: one of "nmb", "inb" or "cutpoints".
 Defaults to "nmb".}
@@ -33,17 +32,13 @@ selected values. Defaults to the median and 95\% interval.}
 The values of the vector are the default names and the names given are the
 desired names in the output.}
 
-\item{show_full_inputs}{A logical. Whether or not to include the inputs used
-for simulation alongside aggregations. (Only applicable for
-\code{predictNMBscreen} objects created using \code{screen_simulation_inputs()})}
-
 \item{...}{Additional, ignored arguments.}
 }
 \value{
 Returns a \code{tibble}.
 }
 \description{
-Create table summaries of simulation objects.
+Create table summaries of \code{predictNMBsim} objects.
 }
 \details{
 Table summaries will be based on the \code{what} argument.
@@ -54,14 +49,13 @@ selected (0, 1).
 }
 \examples{
 
-# perform screen with increasing values of model discimination (sim_auc)
+# perform simulation with do_nmb_sim()
 if (FALSE) {
   get_nmb <- function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
-  sim_screen_obj <- screen_simulation_inputs(
-    n_sims = 50, n_valid = 10000, sim_auc = seq(0.7, 0.9, 0.1), event_rate = 0.1,
-    fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
+  sim_obj <- do_nmb_sim(
+    sample_size = 200, n_sims = 50, n_valid = 10000, sim_auc = 0.7,
+    event_rate = 0.1, fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
   )
-  make_summary_table(sim_screen_obj)
-  make_summary_table(sim_screen_obj$simulations[[1]])
+  make_summary_table(sim_obj)
 }
 }
