+++ man/make_summary_table.predictNMBscreen.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/make_summary_table.R
-\name{make_summary_table}
-\alias{make_summary_table}
-\title{Create table summaries of simulation objects.}
+\name{make_summary_table.predictNMBscreen}
+\alias{make_summary_table.predictNMBscreen}
+\title{Create table summaries of \code{predictNMBscreen} objects.}
 \usage{
-make_summary_table(
+\method{make_summary_table}{predictNMBscreen}(
   x,
   what = c("nmb", "inb", "cutpoints"),
   inb_ref_col = NULL,
@@ -17,7 +17,7 @@ make_summary_table(
 )
 }
 \arguments{
-\item{x}{A \code{predictNMBscreen} or \code{predictNMBsim} object.}
+\item{x}{A \code{predictNMBscreen} object.}
 
 \item{what}{What to summarise: one of "nmb", "inb" or "cutpoints".
 Defaults to "nmb".}
@@ -34,8 +34,7 @@ The values of the vector are the default names and the names given are the
 desired names in the output.}
 
 \item{show_full_inputs}{A logical. Whether or not to include the inputs used
-for simulation alongside aggregations. (Only applicable for
-\code{predictNMBscreen} objects created using \code{screen_simulation_inputs()})}
+for simulation alongside aggregations.}
 
 \item{...}{Additional, ignored arguments.}
 }
@@ -43,7 +42,7 @@ for simulation alongside aggregations. (Only applicable for
 Returns a \code{tibble}.
 }
 \description{
-Create table summaries of simulation objects.
+Create table summaries of \code{predictNMBscreen} objects.
 }
 \details{
 Table summaries will be based on the \code{what} argument.
@@ -62,6 +61,5 @@ if (FALSE) {
     fx_nmb_training = get_nmb, fx_nmb_evaluation = get_nmb
   )
   make_summary_table(sim_screen_obj)
-  make_summary_table(sim_screen_obj$simulations[[1]])
 }
 }
