### 2022-12-20 5a68904
improve tests for make_summary_table outputs
and print() methods too
