### 2022-12-20 dd26967
shorten snapshot names to avoid path length NOTE
(and run styler on plot.R)
