### 2022-12-20 cc07fb0
create snapshots for print methods
