+++ tests/testthat/_snaps/do_nmb_sim.md
@@ -1,0 +1,11 @@
+# print method - works
+
+    predictNMB object
+    
+    Training data sample size:  189
+    Minimum number of events in training sample:  19
+    Evaluation data sample size:  1000
+    Number of simulations:  10
+    Simulated AUC:  0.7
+    Simulated event rate:  0.1
+
