### 2022-12-20 11bc916
remove test for print.predictNMBscreen() snapshot
environment changes for functions in print causing differences from snapshot when tested on new system. Choose to remove this test and instead just remain with the expect_output() alone.
