+++ tests/testthat/_snaps/screen_simulation_inputs.md
@@ -1,21 +1,0 @@
-# print method - works
-
-    predictNMBscreen object
-    
-    There were combinations screened
-    
-    There were multiple inputs ( sim_auc, fx_nmb_training ) that was screened for multiple values:
-    $sim_auc
-    [1] 0.7 0.8 0.9
-    
-    $fx_nmb_training
-    $fx_nmb_training$f1
-    function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
-    <bytecode: 0x000002c3c60d91a0>
-    
-    $fx_nmb_training$f2
-    function() c("TP" = -3, "TN" = 0, "FP" = -1, "FN" = -4)
-    <bytecode: 0x000002c3c60d9600>
-    
-    
-
