### 2022-12-20 823e165
add more testing for get_nmb_sampler()
