+++ R/make_summary_table.R
@@ -23,6 +23,7 @@
 #' selected (0, 1).
 #'
 #' @srrstats {EA2.6} General assessment by assertthat is performed (is.count, is.number) and does not assess or hold expectations on additional class attributes.
+#' @srrstats {EA4.1} Numeric prescision is controlled by the user by modifying the \code{agg_functions} argument and by tibble.
 #' @srrstats {EA5.2} There is no typical onscreen printing from \code{make_summary_table()}. It instead returns a \code{tbl_df/tbl/data.frame} with aggregate values. The default is to round the values but the user is able to overwrite this by using the \code{agg_functions} argument and pass the output to a more general table formatting package like \code{flextable}.
 #' @srrstats {EA5.3} Summary statsistics indicate \code{storage.mode} via tibble.
 #'
