### 2022-12-20 e3cf23c
add tests for do_nmb_sim() and make_summary_table()
