+++ tests/testthat/test-make_summary_table.R
@@ -76,7 +76,7 @@ test_that("predictNMBscreen - make_summary_table works", {
   # test a more complex case
   predictNMBscreen_complex_1 <- make_summary_table(
     obj,
-    rename_vector = c("Treat All" = "all", "Treat None" = "none"),
+    rename_vector = c("Treat All" = "all", "Treat None" = "none", "cost minimising threshold" = "cost minimising"),
     agg_functions = list(mean = mean, min = min, max = max)
   )
   expect_s3_class(predictNMBscreen_complex_1, "tbl")
