### 2022-12-20 #564 @annakrystalli
Ooops! Sorry about that @RWParsons . I think you should be able to run that command so will double check with our both developers. In any case, I'll run it for you. :)
